// frontend/src/types/api.ts
export interface UserData {
  id: number;
  username: string;
  email: string;
  full_name: string;
  role: string;
  role_id: number;
  organization_id?: number | null;
  permissions?: string[] | null;
  organization?: {
    id: number;
    name: string;
  };
}